Split Graph view from Graph controller
The Graph view now has its own file.
